Relocate XLSX frontend modules to app/spreadsheet/
DOCX import lives under app/docs/ and PPTX under app/slides/, so the
XLSX-specific frontend code belongs under app/spreadsheet/, not the
type-agnostic app/documents/ dispatcher. Also order the import menu
XLSX → DOCX → PPTX and add the missing Import PPTX item to the
empty-state dropdown so both menus carry the same options.
